# functions/analytics-svc/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "3.0 - FP0"
  title: Health for Clinical Research
  description: .
basePath: /analytics-svc/api/services/
schemes:
  - http
  - https
produces:
  - application/json
paths:
...
  /cohort:
    x-swagger-router-controller: cohort
    get:
      description: Returns all cohorts in the study
      operationId: getAllCohorts
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: offset
          in: query
          description: offset value for results
          required: false
          type: number
        - name: limit
          in: query
          description: limit value for results
          required: false
          type: number
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CohortListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      description: Sends mriquery to patient analytics service to get list of patient from mriquery, extracts all patient ids, then creates cohort definition and cohort entries in database
      operationId: createCohort
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: cohort
          schema:
            type: object
            required:
              - datasetId
              - mriquery
              - description
            properties:
              datasetId:
                type: string
                description: Id of dataset
              mriquery:
                type: string
                description: Filtercard information in bookmark form. This is a binary string encoded in base64
              description:
                type: string
                description: Cohort Description
              definitionTypeConceptId:
                type: string
                description: Cohort Definition Type Concept Id
              subjectConceptId: 
                type: number
                description: Cohort Definition Subject Concept Id
              syntax: 
                type: string
                description: Cohort Definition Syntax
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    delete:
      description: Deletes a cohort with corresponding incoming cohortId
      operationId: deleteCohort
      produces:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: cohortId
          in: query
          type: number
          description: ID of cohort to be deleted from database
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /cohort/{filterColumn}/{filterValue}:
    x-swagger-router-controller: cohort
    get:
      description: Returns list of cohorts based on input filter column and value
      operationId: getFilteredCohorts
      produces:
        - application/json
      parameters:
        - name: filterColumn
          in: path
          description: Filter column that maps to a column name in cohort definition table
          required: true
          type: string
        - name: filterValue
          in: path
          description: Filter value to filter based on filter column
          required: true
          type: string      
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: offset
          in: query
          description: offset value for results
          required: false
          type: number
        - name: limit
          in: query
          description: limit value for results
          required: false
          type: number
      responses:
        "200":
          description: success
          schema:
            $ref: "#/definitions/CohortListResponse"
        default: 
          description: Error
          schema:
            $ref: "#/definitions/Error"
  /cohort-definition:
    x-swagger-router-controller: cohort
    get:
      description: Get a cohort definition via cohort definition id
      operationId: getCohortDefinition
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - name: datasetId
          in: query
          description: Id of dataset
          required: true
          type: string
        - name: cohortDefinitionId
          in: query
          description: Id of cohort definition
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    post:
      description: Saves cohort definition to db
      operationId: createCohortDefinition
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: cohortDefinition
          schema:
            type: object
            required:
              - datasetId
              - name
              - description
            properties:
              datasetId:
                type: string
                description: Id of dataset
              name:
                type: string
                description: Cohort Name
              description:
                type: string
                description: Cohort Description
              definitionTypeConceptId:
                type: string
                description: Cohort Definition Type Concept Id
              subjectConceptId: 
                type: number
                description: Cohort Definition Subject Concept Id
              syntax: 
                type: string
                description: Cohort Definition Syntax
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
    put:
      description: Updates a cohort definition
      operationId: updateCohortDefinition
      produces:
        - application/json
      consumes:
        - application/json
      parameters:
        - in: body
          name: cohortDefinition
          schema:
            type: object
            required:
              - datasetId
              - cohortDefinitionId
            properties:
              datasetId:
                type: string
                description: Id of dataset
              cohortDefinitionId:
                type: number
                description: Cohort Definition Id
              name:
                type: string
                description: Cohort Definition Name
              description:
                type: string
                description: Cohort Definition Description
              definitionTypeConceptId:
                type: number
                description: Cohort Definition Definition Type Concept Id
              syntax:
                type: string
                description: Cohort Definition Syntax
              subjectConceptId:
                type: number
                description: Cohort Definition Subject Concept Id
      responses:
        "200":
          description: Success
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
...
  CohortResponse:
    type: object
    description: Cohort belonging to the study
    required: [id, patientIds, name, description, creationTimestamp]
    properties:
      id:
        type: string
        description: Unique identifies for a cohort
      patientIds:
        type: array
        description: Array of patient IDs belonging to cohort
        items:
          type: string
      name:
        type: string
        description: Name of cohort
      description:
        type: string
        description: Description of cohort
      creationTimestamp:
        type: string
        description: Indicates when the cohort was created
  CohortListResponse:
    type: array
    description: Array of cohorts
    items:
      $ref: "#/definitions/CohortResponse"
...